requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
iniconfig==2.0.0
packaging==24.0
pip==23.2.1
pluggy==1.5.0
pytest==8.2.0
requests==2.31.0
setuptools==68.2.0
urllib3==2.2.1
wheel==0.41.2
